Port plugin to Elasticsearch 5.1.1
Quite a lot of things changed between 2.x and 5.x: not only the code
itself had to be changed, but the build needed to be upgraded as well.
Unfortunately there was no straightforward way to do these things one
at a time so this commit does a bit more things that is comfortable. In
the end the things are relatively straightforward:

  - change code to use APIs of ES 5.1.1
  - convert tests to leverage ESTestCase
  - convert build to use Gradle along with 'elasticsearch.esplugin'
  - make a bunch of changes required by 'elasticsearch.esplugin'
  - relicense code to use Apache 2 license